Migrate baseballgame class into index.js
