+++ README.md
@@ -19,6 +19,7 @@
   - 볼의 개수를 세기.
 - 재시작 버튼을 눌렀을 때 reload.
   - 정답을 맞힌 뒤 재입력 방지.
+  - 재시작 버튼에 game-restart-button 아이디 부여.
 
 ## ✔️ 이슈
