+++ index.css
@@ -2,6 +2,10 @@ body {
   font-family: "Noto Sans KR", sans-serif;
 }
 
+#answer {
+  font-weight: 500;
+}
+
 #game-restart-button {
   margin-left: 10px;
 }
